requirements.txt
Flask==3.1.0
pytesseract==0.3.13
Pillow==10.0.0
